fixing launching upload image dialog
