+++ app/src/main/java/com/example/thesimplesocialapp/AddAccountActivity.java
@@ -1,5 +1,7 @@
 package com.example.thesimplesocialapp;
 
+import android.app.Activity;
+import android.content.Intent;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.os.Bundle;
@@ -9,6 +11,11 @@ import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
 
+import androidx.activity.result.ActivityResult;
+import androidx.activity.result.ActivityResultCallback;
+import androidx.activity.result.ActivityResultLauncher;
+import androidx.activity.result.contract.ActivityResultContract;
+import androidx.activity.result.contract.ActivityResultContracts;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
@@ -16,6 +23,8 @@ import androidx.core.view.WindowInsetsCompat;
 
 public class AddAccountActivity extends AppCompatActivity implements IRegisImagePath{
     private String regis_image_path;
+    ActivityResultLauncher<Intent> imageChooserLauncher;
+    // ======================= DELETE ===========================
     public void setRegis_image_path(String regis_image_path) {
         this.regis_image_path = regis_image_path;
     }
@@ -34,6 +43,22 @@ public class AddAccountActivity extends AppCompatActivity implements IRegisImage
         Button regis_btn = findViewById(R.id.regis_btn);
         UploadImageDialogFragment upload_regis_image_dialog = new UploadImageDialogFragment();
 
+        // ======================= DELETE ===========================
+        imageChooserLauncher = registerForActivityResult(new ActivityResultContracts.StartActivityForResult(),
+                new ActivityResultCallback<ActivityResult>() {
+                    @Override
+                    public void onActivityResult(ActivityResult o) {
+                        String filePath;
+                        Log.i("h", "hello world");
+//                        mCallback.returnDataFromUploadImage("hello world");
+                        if (o.getResultCode() == Activity.RESULT_OK) {
+//                        Uri data = o.getData().getData();
+//                        filePath = getPath(data);
+//                        Log.i("image path", filePath);
+                        }
+                    }
+                });
+
         // init db
         try{
             db_LocalCred database = new db_LocalCred(AddAccountActivity.this);
@@ -46,7 +71,19 @@ public class AddAccountActivity extends AppCompatActivity implements IRegisImage
         regis_btn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
+                Intent galleryIntent = new Intent(Intent.ACTION_PICK, android.provider.MediaStore.Images.Media.EXTERNAL_CONTENT_URI);
+                ActivityResultLauncher<Intent> dialog = registerForActivityResult(new ActivityResultContracts.StartActivityForResult(),
+                        new ActivityResultCallback<ActivityResult>() {
+                            @Override
+                            public void onActivityResult(ActivityResult o) {
+
+                            }
+                        }
+                );
                 upload_regis_image_dialog.show(getSupportFragmentManager(),"getImageUpload");
+
+                // ======================= DELETE ===========================
+                imageChooserLauncher.launch(galleryIntent);
             }
         });
 
@@ -56,8 +93,8 @@ public class AddAccountActivity extends AppCompatActivity implements IRegisImage
                 finish();
             }
         });
-
     }
+
 
     private Cursor getEvents(SQLiteDatabase db){
         String SELECT = ILocalCred.TABLE_NAME;
