+++ app/src/main/java/com/example/thesimplesocialapp/UploadImageDialogFragment.java
@@ -26,18 +26,6 @@ import androidx.loader.content.CursorLoader;
 
 public class UploadImageDialogFragment extends DialogFragment {
     private static final int PICK_IMAGE_REQUEST = 1;
-    private IRegisImagePath mCallback;
-
-    @Override
-    public void onAttach(@NonNull Context context) {
-        super.onAttach(context);
-        try {
-            mCallback = (IRegisImagePath) context;
-        }
-        catch (ClassCastException e) {
-            Log.d("Register image upload", "Activity doesn't implement the needed interface");
-        }
-    }
 
     @NonNull
     @Override
@@ -45,9 +33,8 @@ public class UploadImageDialogFragment extends DialogFragment {
         // Use the Builder class for convenient dialog construction.
         AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
         builder.setMessage("Please upload an image to use as a profile picture.")
-                .setPositiveButton("Upload", new DialogInterface.OnClickListener() {
+                .setPositiveButton("Proceed", new DialogInterface.OnClickListener() {
                     public void onClick(DialogInterface dialog, int id) {
-                        showFileChooser();
                     }
                 })
                 .setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
@@ -57,38 +44,4 @@ public class UploadImageDialogFragment extends DialogFragment {
         // Create the AlertDialog object and return it.
         return builder.create();
     }
-    private void showFileChooser() {
-        Intent galleryIntent = new Intent(Intent.ACTION_PICK, android.provider.MediaStore.Images.Media.EXTERNAL_CONTENT_URI);
-        // Start the Intent
-        imageChooserLauncher.launch(galleryIntent);
-    }
-    ActivityResultLauncher<Intent> imageChooserLauncher = registerForActivityResult(
-            new ActivityResultContracts.StartActivityForResult(),
-            new ActivityResultCallback<ActivityResult>() {
-                @Override
-                public void onActivityResult(ActivityResult o) {
-                    String filePath;
-                    if (o.getResultCode() == Activity.RESULT_OK) {
-                        Uri data = o.getData().getData();
-                        filePath = getPath(data);
-                        Log.i("image path", filePath);
-                        mCallback.returnDataFromUploadImage("hello world");
-                    }
-                }
-            });
-    // upload image
-    private void uploadImage(String path){
-    }
-    // Get Path of selected image
-    private String getPath(Uri contentUri) {
-        String[] proj = { MediaStore.Images.Media.DATA };
-        CursorLoader loader = new CursorLoader(getContext(),contentUri, proj, null, null, null);
-        Cursor cursor = loader.loadInBackground();
-        int column_index = cursor.getColumnIndexOrThrow(MediaStore.Images.Media.DATA);
-        cursor.moveToFirst();
-        String result = cursor.getString(column_index);
-        cursor.close();
-        return result;
-    }
-
 }
